Refactor server/server.go to accept generic requests
The server was previously written to only accept HTTP requests, and
as a result, the server functions take `http.Request`s and
`http.ResponseWriter`s as arguments. This patch reworks the server so
that its functions use more generic `request` and `response` structs.
Separate functions specific to HTTP translate to and from HTTP
constructs and the generic ones.

This is needed for a future patch which will integrate gRPC into the
server as an alternate access method.