Implement delta watches for Set documents in the server
For some clients, it might be useful to be able to receive a diff
through the watch API whenever something changes, but for semantics
reasons, watches only support sending diffs if the document being
watched is a Set. There is one exception to this: it is valid to set a
watch on a Null document, which will send updates whenever something
changes as if it is an empty Set.

The reasoning for this is that sometimes the data backing the document
may not be defined yet. While the transition from Null -> Set is
supported, if the document ever changes away from Null though, it may
not change back. That is, Null -> Set -> Null will cause the watch to
error.